Move serve hosted page under mcp_install_page
