+++ server/internal/mcpinstallpage/config_snippet.json.tmpl
