+++ server/internal/mcpinstallpage/hosted_page.html.tmpl
